Update ios-add-files-via-xcode.png
The ios-add-files-via-xcode.png file shows the right click on the top Runner instead of on the Runner inside the Runner. If you try to add the GoogleService-Info.plist as it shows in the image it add the file in a wrong place and it dont work. I share a link with a new image.